fix(editor): accept fractional frame rates end to end
Real-user export attempt on real camera footage (23.976 fps NTSC rate) blew
up at editor/start: EditorDocument declared fps as int, so every V2 candidate
cut from fractional-rate sources was rejected at the editor boundary — while
the engine itself handles those rates fine. pov never tripped it (phone
footage, flat 30).

- EditorDocument.fps: int -> float (bounds kept)
- renderer2: fps was int()-truncated — 23.976 would render at 23, a 4% speed
  error. Now carried fractionally (ffmpeg fps/r filters accept decimals).
- Test mirrors the exact failing path: 23.976 candidate ->
  document_from_candidate -> EditorDocument.

Full backend suite green.

+++ render-backend/app/product_editor.py
@@ -28,7 +28,10 @@ class EditorDocument(BaseModel):
     candidateRunId: UUID
     width: int = Field(ge=16, le=7680)
     height: int = Field(ge=16, le=4320)
-    fps: int = Field(ge=1, le=120)
+    # float: real footage runs at fractional rates (23.976, 29.97). Declaring
+    # int made editor/start reject every V2 candidate cut from NTSC-rate
+    # sources while the engine itself handled them fine.
+    fps: float = Field(ge=1, le=120)
     duration: float = Field(gt=0, le=MAX_DURATION_SECONDS)
     tracks: list[dict]
     sourceAssetIds: list[UUID] = Field(min_length=1)
